move agate_helper into common
