+++ core/dbt/common/clients/agate_helper.py
